Fixes for renamed process library
